Change example class names
Use completely different terms so that there's no
confusion between using an id that is the singular
of the plural name used in the classes example.
There's a chance for confusion.

Try to use more obviously multiple yet singular
classes e.g. important-topic(s) are likely to be
present as are helpful-hint(s).